+++ src/meshProvider.ts
@@ -124,6 +124,12 @@ export class MeshViewProvider
       pointColor: config.get("pointColor", "#cc0000"),
       wireframeColor: config.get("wireframeColor", "#0000ff"),
       fogDensity: config.get("fogDensity", 0.01),
+      coordinateSystem: config.get("coordinateSystem", "opengl"),
+      customCoordinateSystem: config.get("customCoordinateSystem", {
+        rightAxis: "+x",
+        upAxis: "+y", 
+        forwardAxis: "+z"
+      }),
     };
     return `<meta id="vscode-3dviewer-data" data-settings="${JSON.stringify(
       initialData